moved contents of `Pathish` to `PathOps`
